Parameter StatLogger.S

val id : string

Unique logger id, must be the same between all runs

type stat

The type of stats being saved. For example:

A stat aggregate int t or float t (in which case combine is just fun _ -> concat)

A record of such stat aggregates (combine is then the field-wise concat)

A list of values (combine is then List.concat)

val combine : string -> stat -> stat -> stat